Fixed #30385 -- Restored SearchVector(config) immutability.
Regression in 8e05bf3adf628dd0306eb5385ea794b483867206.

The usage of CONCAT to allow SearchVector to deal with non-text fields
made the generated expression non-IMMUTABLE which prevents a functional
index to be created for it.

Using a combination of COALESCE and ::text makes sure the expression
preserves its immutability.

Refs #29582. Thanks Andrew Brown for the report, Nick Pope for the
review.