Fixed #33700 -- Skipped extra resolution for successful requests not ending with /.
By moving a should_redirect_with_slash call out of an if block, commit
7bce8dc108243a69f81f456b38fe1a8a56b1ffbd negated the performance fix
of commit fc6bcd3ff18d89de7542872f08d81d47da797d2e (#24720).
Meanwhile, the logging issue #26293 that it targeted was subsequently
fixed more fully by commit 926c6439e646e852fdade29e6adc43b54d4d73f7
(#26504), so it is no longer needed.  This effectively reverts it.

This speeds up successful requests not ending with / when APPEND_SLASH
is enabled (the default, and still useful in projects with a mix of
URLs with and without trailing /).  The amount of speedup varies from
about 5% in a typical project to nearly 50% on a benchmark with many
routes.

Signed-off-by: Anders Kaseorg <andersk@mit.edu>