Fixed #33992 -- Fixed queryset crash when aggregating over a group containing Exists.
A more in-depth solution is likely to make sure that we always GROUP BY
selected annotations or revisit how we use Query.exists() in the Exists
expression but that requires extra work that isn't suitable for a
backport.

Regression in 5d652bba1f1083228a34fe1a3773a6d450dab864.

Thanks Fernando Flores Villaça for the report.