Fixed #34024 -- Fixed crash when aggregating querysets with Q objects annotations.
This reverts eefd4d8b9e283aee026dd660418b225eebed5c2c.

It was reasonable to assume it was unnecessary code as there were
no failing tests upon its removal. This commit adds the necessary
regression tests for the failing condition identified in #34024
alongside the original tests added in the PR for which
WhereNode.is_summary was introduced.