Fixed #34612 -- Fixed QuerySet.only() crash on reverse relationships.
Regression in 21364d4b79f2163b0ef21e47e0865e7881a2eebf.

Thanks Ian Cubitt for the report.

This also corrected test_inheritance_deferred2() test which was
previously properly defined and marked as an expected failure but was
then wrongly adjusted to mask the lack of support for per-alias
deferral that was fixed by #21204.